feat(arcs): move arc hover detection from core to arcs package
