+++ packages/core/src/lib/interactivity/detect.js
@@ -1,12 +1,3 @@
-/*
- * This file is part of the nivo project.
- *
- * Copyright 2016-present, Raphaël Benitte.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 /**
  * Computes distance between two points.
  *
@@ -54,42 +45,3 @@ export const getAngle = (x1, y1, x2, y2) => {
  */
 export const isCursorInRect = (x, y, width, height, cursorX, cursorY) =>
     x <= cursorX && cursorX <= x + width && y <= cursorY && cursorY <= y + height
-
-/**
- * Check if cursor is in given ring.
- *
- * @param {number} centerX
- * @param {number} centerY
- * @param {number} radius
- * @param {number} innerRadius
- * @param {number} cursorX
- * @param {number} cursorY
- * @return {boolean}
- */
-export const isCursorInRing = (centerX, centerY, radius, innerRadius, cursorX, cursorY) => {
-    const distance = getDistance(cursorX, cursorY, centerX, centerY)
-
-    return distance < radius && distance > innerRadius
-}
-
-/**
- * Search for an arc being under cursor.
- *
- * @param {number}         centerX
- * @param {number}         centerY
- * @param {number}         radius
- * @param {number}         innerRadius
- * @param {Array.<Object>} arcs
- * @param {number}         cursorX
- * @param {number}         cursorY
- * @return {*}
- */
-export const getHoveredArc = (centerX, centerY, radius, innerRadius, arcs, cursorX, cursorY) => {
-    if (!isCursorInRing(centerX, centerY, radius, innerRadius, cursorX, cursorY)) return null
-
-    const cursorAngle = getAngle(cursorX, cursorY, centerX, centerY)
-
-    return arcs.find(
-        ({ startAngle, endAngle }) => cursorAngle >= startAngle && cursorAngle < endAngle
-    )
-}
